Set default focus to pick color button. Closes #4
Setting the pick color button as the default focused
element would be a more sensible default as opposed
to the current automatically picked focus element,
which seems to default to the color format text entry.
The user would probably want to start picking as soon
as the app is opened and focusing the pick button would
enable the user to press Enter/Space to immediately
start picking.