Implement color picking using portal
Use freedesktop color picking portal to implement color picking.
This takes away the in window preview functionality for now,
but there is no straight forward way to implement that in
gtk4 + wayland anyway. The preview should be left to the portal
implementation. This greatly simplifies the code and improves
compatibility as it should work anywhere where the portal is
implemented. Should help towards #7